fix: decorator jsx 태그 안되는이슈 확장자 ts -> tsx로 고쳐서 해결 및 헤더 잘보이도록 배경 적용
